feat(workflow): version and tag update (#6)
## feat(mirror): remove `v` from version
    
As mentioned in https://github.com/astral-sh/uv-pre-commit/issues/5,
`uv` does not use `v` in its version tagging. Considering that this repo
is primarily a mirror of the `uv` repo, it should match.

## feat(workflow): auto release on tag creation
    
Although this functionality is not required, some users might be
confused if the tag and release version are not the same. This in
reference to https://github.com/astral-sh/uv-pre-commit/issues/4
    
Additionally, the body of the release points to the uv release