fix: auto release on mirror updates (#8)
As mentioned in https://github.com/astral-sh/ruff-pre-commit/issues/82 /
https://github.com/astral-sh/ruff-pre-commit/pull/83, the auto release
function doesn't work. This fixes that by checking for changes and
setting a flag. If the flag is set, it will also create a new release
through gh cli.

Another note is that the `actions/setup-python` is updated from `v4` →
`v5`
 
This is being done now because it has been working well for the
ruff-pre-commit repo (I think?)