Add `--no-active` to uv-sync hook (#43)
Add `--no-active` option to `uv-sync` to ignore the pre-commit virtual
env. Closes #36.

As far as I can tell, the other hooks don't support/need this option,
but would like that confirmed.

### Testing

To test this PR, add the following to `.pre-commit-config.yaml`

```
  - repo: https://github.com/danielhollas/uv-pre-commit    
    rev: fc377b7
    hooks:    
    - id: uv-sync                                                                                                                         
      verbose: true
      stages: [pre-commit]
```

and run

```console
pre-commit run -a uv-sync
```

No warning should be printed.